+++ src/main/scala/ot/scalaotl/commands/OTLEventstats.scala
@@ -2,23 +2,39 @@ package ot.scalaotl
 package commands
 
 import org.apache.spark.sql.DataFrame
+import ot.scalaotl.commands.commonconstructions.{StatsContext, StatsTransformer}
 import ot.scalaotl.extensions.StringExt._
 import ot.scalaotl.parsers.StatsParser
 
 class OTLEventstats(sq: SimpleQuery) extends OTLBaseCommand(sq, _seps = Set("by")) with StatsParser {
   val requiredKeywords = Set.empty[String]
   val optionalKeywords = Set.empty[String]
-  val transformer = new OTLStats(SimpleQuery(args))
-  override val fieldsUsed: List[String] = transformer.fieldsUsed
-  override val fieldsGenerated: List[String] = transformer.fieldsGenerated.map(_.stripBackticks())
+  val transformer = new StatsTransformer(sq, StatsContext(spark, returns, positionalsMap, getKeyword("timeCol").getOrElse("_time")))
+  override val fieldsUsed: List[String] = getFieldsUsed(returns) ++ getPositionalFieldUsed(positionals)
+  override val fieldsGenerated: List[String] = getFieldsGenerated(returns).map(_.stripBackticks())
 
   override def transform(_df: DataFrame): DataFrame = {
-    val _df_out = transformer.transform(_df)
-    val res = positionalsMap.get("by") match {
-      case Some(Positional("by", List())) => _df.drop(fieldsGenerated: _*).crossJoin(_df_out)
-      case Some(Positional("by", byList)) => _df.drop(fieldsGenerated: _*).join(_df_out, byList.map(_.stripBackticks()))
-      case _ => _df_out
+    val outDf = transformer.transform(_df)
+    val workDf = _df.drop(fieldsGenerated: _*)
+    val eventstatsDf = positionalsMap.get("by") match {
+      case Some(Positional("by", List())) => workDf.crossJoin(outDf)
+      case Some(Positional("by", byList)) => {
+        val workByList = byList.map(_.stripBackticks())
+        val colItem = workDf(workByList.head) <=> outDf(workByList.head)
+        val cols = workByList.tail.foldLeft(colItem) {
+          (col, p) => col && workDf(p) <=> outDf(p)
+        }
+        val outColItem = outDf(workByList.head)
+        val outCols = workByList.tail.foldLeft(outColItem) {
+          (outCol, p) => outCol && outDf(p)
+        }
+        workDf.join(outDf, cols).drop(outCols)
+      }
+      case _ => outDf
     }
-    res
+    if (eventstatsDf.columns.contains("_time"))
+      eventstatsDf.sort("_time")
+    else
+      eventstatsDf
   }
 }